Use uint64_t to represent num_retired_cids
This patch changes the datatype of num_retired_cids to uint64_t,
which is the standard datatype in quicly for variable-length values.

In the current implementation, maximum number of CIDs is limited to
MAX_PATH_ID (=255), so in that sense uint8_t could be enough.
However, if we are to use uint8_t, we would need to carefully monitor
the change in MAX_PATH_ID value range, which is a heavy burden on
maintenance.

+++ include/quicly.h
@@ -475,7 +475,7 @@ struct _st_quicly_conn_public_t {
          */
         quicly_cid_t offered_cid;
         struct st_quicly_conn_streamgroup_state_t bidi, uni;
-        uint32_t num_retired_cids;
+        uint64_t num_retired_cids;
     } host;
     struct {
         /**
